Reduce ability to create invalid RDS template
Implement a complex mechanism for reducing the chance of creating JSON
for an RDS DBInstance that is invalid.  My hope is that the complexity
equals and is not greater than the complexity inherent in the
restrictions around RDS creation.  There is still one run-time check
around storage encryption since specifying it is only valid on new
instances created within VPCs.

There are restrictions on the values of AllocatedStorage when using
Iops but often these values are given as parameter and therefore must
accept Token[Int].  I currently used Either to take Int or Token[Int]
but at some point it would be good to allow one to reach through the
Token to see if the Int is there too test.

Tests were added to cover all of the above.

A couple other minor changes:
-   Clean up build.sbt
-   Implement and use EnumFormat for all enumerations JSON serialization