Make LogDestination.DestinationPolicy take a JsonString
It clearly states in the docs that this is a String argument,
and we should not force people to make an IAM Policy and attach
it to some random group or user just so we can use it in
this field.

I also don't know what is up with those private constructors
without object constructors.